drm/i915: Track which pipe is in CxSR mode
Before enabling a second pipe we have to turn off CxSR and wait for
the first pipe to exit CxSR mode. Otherwise the first pipe can still
be using the CxSR FIFO configuration when the second pipe comes up,
and that can lead to underruns.

The simplest way to do this is to just track the "sr_pipe" and
add do the vblank wait on it when SR is getting disabled by
the watermark programming done for the second pipe.

v2: .disable_cxsr() hook

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>