drm/sti: Try to fix up the tvout possible clones
The current possible_clones setup doesn't look sensible. I'm assuming
the 0 and 1 are supposed to refer to the indexes of the hdmi and hda
encoders? So it kinda looks like we want hda+hdmi cloning, but then
dvo also claims to be cloneable with hdmi, but hdmi won't recipricate.

So let's assume we just want hdmi+hda and nothing more and populate the
possible_clones to match. The docs say the encoder itself should always
be included in the possible_clones bitmask so we'll throw it in there
as well. For hda we'll leave possible_clones==0 and we'll add a global
fallback for that case to add the encoder itself to the bitmask since
that makes life easier for most drivers.

Cc: Benjamin Gaignard <benjamin.gaignard@linaro.org>
Cc: Vincent Abriou <vincent.abriou@st.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>