drm/i915: Filter out modes that don't match the fixed mode vrefresh
We only ever drive the panel with the fixed mode, hence we don't
want to advertize any modes that have a different vertical refresh rate.

We tried to allow a second lower clocked mode to used for eDP but
that was reverted in commit d93fa1b47b8f ("Revert "drm/i915/edp:
Allow alternate fixed mode for eDP if available.""). To do that properly
I think we should probably just keep all (or just some?) of the probed
modes around (presumably all of them actually work if the panel
advertizes them), and we'd just pick the closest match based on the
user mode.

For now we don't have any of that so we shouldn't lie to the user
that they can drive the panel at different refresh rate. Note that
we still don't reject any user mode with bad refresh rate. That's
going to be step two.

TODO: Should we allow for a larger error margin?

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>