drm/i915: Pass the new crtc_state to ->disable_plane()
We're going to need access to the new crtc state in ->disable_plane()
for SKL+ wm/ddb programming and pre-skl pipe gamma/csc control. Pass
the crtc state down.

We'll also try to make intel_crtc_disable_planes() drtr as much as
it's possible. The fact that we don't have a separate crtc state
for the disabled state when we're going to re-enable the crtc later
means we might end up poking at a few extra planes in there. But
that's harmless. I suppose one migth argue that we wouldn't have to
care about proper ddb/wm/csc/gamma if the pipe is going to permanently
disable anyway, but the state checker probably cares so we should try
our best to make sure everything is programmed correctly even in that
case.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>