drm/i915: Don't use split gamma when we don't have to
Using the split gamma mode when we don't have to has the annoying
requirement of loading a linear LUT to the unused half. Instead
let's make life simpler by switching to the 10bit gamma mode
and duplicating each entry.

This also allows us to load the software gamma LUT into the
hardware degamma LUT, thus removing most of the bad failure
modes. The user can now use the gamma LUT freely even if we
are going to do YCbCr or RGB limited range output. Only
when the user asks for degamma/ctm do we start to hit
combinations that can't be supported. Sadly GLK+ lost
all this flexibility with its dedicated degamma LUT.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>