drm/i915: Add more error checks to color_check()
We have several conflicting combinations of csc and
gamma stuff which we're not rejecting. Do so.

This will probably "break" somehting though :( We should
implement at least some of the ideas listed in the FIXMEs to
avoid users getting failures.

My main concerns are the RGB limited range + gamma LUT,
and YCbCr + gamma LUT cases. We enable both the RGB
limited range and YCbCr output automagically, so the
user doesn't really get to choose between working gamma
vs. not.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>