drm/i915: Turn dram_info.num_channels into a bitmask
We want to know out which channels are actually occupied so that
later on we can read the memory timings from the right registers.
To that end convert num_channels into a bitmask.

v2: Don't WARN if we can't read the dunit regs on bxt/glk
v3: num_channels is now unused

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>