drm/i915: Expose margin properties on ilk+ HDMI
Add support for underscan by exposing the margin properties on
ilk+ HDMI ports. This can be useful with silly TVs that won't
let you disable overscanning.

We limit this to ilk+ only because the gmch style panel fitter
doesn't really have the flexibility we need.

TODO: DP and/or lspcon?

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>