drm/i915: Try to probe DP++ dongles on DP++ downstream facing ports
In order to get accurate TMDS clocks limits for the DP++ ports we
can try to probe the DP dual mode adapter registers. Unfortunately
I've not yet seen a DP->DP++ branch device that would actually
forward these i2c accesses to the dual mode dongle downstream.
But we don't lose much by trying and maybe it'll work on some
branch devices, if not now then maybe in the future.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>