drm/i915: Use vblank workers for gamma updates
The pipe gamma registers are single buffered so they should only
be updated during the vblank to avoid screen tearing. In fact they
really should only be updated between start of vblank and frame
start because that is the only time the pipe is guaranteed to be
empty. Already at frame start the pipe begins to fill up with
data for the next frame.

Unfortunately frame start happens ~1-4 scanlines after the start
of vblank which in practie doesn't always leave us enough time to
finish the gamma update in time (gamma LUTs can be several KiB of
data we have to bash into the registers). However we must try our
best and so we'll add a vblank work for each pipe from where we
can do the gamma update.

Even though the vblank worker is running on a high prority thread
we still have to contend with C-states. If the CPU happens be in
a deep C-state when the vblank interrupt arrives even the irq
handler gets delayed massively (I've observed dozens of scanlines
worth of latency). To avoid that problem we'll use the qos mechanism
to keep the CPU awake while the vblank work is scheduled.

With all this hooked up we can finally to almost atomic gamma
updates. It even works across several pipes from the same atomic
commit which previously was a total fail because we did the
gamma updates for each pipe serially. In practice we still miss
the frame start easily and so ~100 pixels of the new frame
generally make it through with the old gamma. I don't see any
way to avoid that so we'll just have to live with it. Also it's
possible that high system load etc. could still cause us blow
past the end of vblank, but not much we can do there either
except rewrite the scheduler.

In the future the DSB should take over this responsibility
which will hopefully avoid some of these issues.

v2: Add missing intel_atomic_state fwd declaration
v3: Clean up properly when not scheduling the worker

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>