drm/i915: Perform single buffer plane register updates from vblank worker
Like we did for pipe gamma let's move the single buffered plane
register updates to occur from a vblank work.

As the vblank works are a bit on the expensive side (with the qos and
everything) let's try to do this only when we absolutely have to. If
we know that the registers we're about to update don't have any
effect on the old frame we can just directly write the registers
while the old frame is still being scanned out.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>