drm/i915: Call skl_detach_scalers() only from a single place
The code is silly and calls skl_detach_scaler() from both sides
of an if-else. Just move the call to happen before the if-else.

We probably still want to keep this as close to the pfit/plane
programming as reasonably possible so move it just before the
intel_pipe_fastset() call.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>