drm/i915: Check pipe scaler scaling factors
Make sure we don't exceed the maximum scaling factors of scalers.
For plane scaling we've already done a preliminary check but that
check doesn't account for the difference between HQ vs. DYN scaler
modes (since we do the check before we know whether HQ mode can
be used or not). For pfit use cases this will be the only scaling
factor check we do.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>