drm/i915: Split crtc and plane state to uapi and internal copies
Split the crtc and plane states into uapi and internal copies so
that we can freely modify the internal copy without affecting
what userspace sees as the current state. This is requires for
stealing pipes to enable the bigjoiner 2pipes->1port use case.

I've simply duplicated the drm_*_state base member into 'uapi'
and 'base' (the internal one) copies. I did it this way so that
we don't have to modify all the code that accesses the base
member (of which there is a lot). We do the uapi->base copy
at specific points in the sequence to avoid coupling the entire
driver to the bigjoiner logic.

No bijoiner speciifc stuff is present here yet. We need to do the
state copy for that in a slightly different way as the duplicated
states need to be adjusted.

base.commit, base.fence, and base.state are declared verboten so
that we don't get into a horrible mess with the core/helpers.
Should probably sprinkle a few WARNs around to make sure these
stay NULL always. Ideally we'd remove such things from the base
member entirely, but that would mean we'd have to hand roll the
uapi->base copy entirely. Maybe do that as a followup if this
approach works out...

Caveat: I was too chicken to boot this yet...

Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Cc: Matt Roper <matthew.d.roper@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>