drm/i915: Add missing commans to dbuf tables
The preferred style is to sprinkle commas after each array and
structure intialization, whether or not it happens to be the
last element/member. This leads to much prettier diffs if/when
new elements/members get added to the end of the initialization.
We're not bound by some ancient silly mandate to omit the final
comma.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>