drm/i915: Decouple cdclk from intel_modeset_checks()
Now that cdclk is using the global state thing we can decouple it from
any_ms and intel_modeset_checks(). That is, we just got ahead and
recompute cdclk without the rest of the modeset checks if and when
necessary. The programming side also has to pulled out from underneath
the state->modeset checks, which is fine since the cdclk code does the
global state serialization and forces everything to go sequentially
regardless of which workqueue handles the commit.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>