drm/i915: Reset force_min_cdclk lazily
The current way if forcing cdclk up and immediately back down around
audio get/put_power() calls is causing tons of ugly display blinking.
To combat this let's try to leave the cdclk in the up state after
put_power(), and leave it up to the next commit to push it back
down if it's not needed for any other reason.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>